Create MovableLayerScript
The var statement at beginning of script defines two variables, x and y, that will store current position of container. The pos function is called by event handlers for all four of the movement buttons. Parameters of the pos() function, dx and dy, tell the script how container should move: If dx is negative, a number is subtracted from x, moving the container to the left. If dx is positive, a number is added to x, moving the container to the right. Similarly, dy indicates whether to move up or down.

The pos() function begins by making sure getElementById() function is supported, so it won't attempt to run in older browsers. It then multiplies dx and dy by 30 (to make movement more obvious) and applies them to x and y. Finally, it sets the top and left properties to the new position (including the "px" to indicate unit of measurement), thus moving the container layer.

Two more functions, hideSquare() and showSquare(), hide or show the container by setting its display property to "none" (hidden) or "block" (shown).